requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.20.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
jupyter>=1.0.0
ipywidgets>=8.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
# Note: For GPU acceleration, replace faiss-cpu with faiss-gpu
